requirements.txt
Flask==2.2.5
Werkzeug==2.3.7
prometheus-client==0.16.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-flask==0.48b0
pytest==7.4.2
requests==2.31.0
